Refactor imports in multiple files to remove relayInit and enhance component standalone compatibility; update EventComponent and ProfileActionsComponent to utilize Angular imports.
